added createAsyncThunk for handling async actions, start service
